fix: resolve RV64 test failures in algebra circuit tests
Fix three issues blocking algebra circuit unit tests:

1. Test infrastructure (cpu.rs, cuda.rs): write_heap and
   write_heap_pointer_default were hardcoded to RV32_REGISTER_NUM_LIMBS (4
   bytes). Updated to RV64_REGISTER_NUM_LIMBS (8 bytes) so register
   writes match what RV64 adapters expect.

2. Adapter record alignment (vec_heap.rs, vec_heap_branch.rs, eq_mod.rs):
   Rv64*AdapterRecord structs had u64 fields requiring 8-byte alignment,
   but MatrixRecordArena only guarantees 4-byte alignment (backed by
   Vec<BabyBear> = Vec<u32>). Changed u64 fields to [u8; 8] to reduce
   struct alignment to 4 bytes while preserving the same size.

3. Algebra test fixes (modular_chip/tests.rs, fp2_chip/tests.rs):
   write_ptr_reg now expects u64; added .into() conversions. fp2 tests
   used u32::to_le_bytes for register writes; changed to u64::to_le_bytes.

+++ extensions/algebra/circuit/src/fp2_chip/tests.rs
@@ -199,17 +199,17 @@ fn set_and_execute_fp2<const BLOCKS: usize, const BLOCK_SIZE: usize, const NUM_L
     tester.write::<RV64_REGISTER_NUM_LIMBS>(
         ptr_as,
         rs1_ptr,
-        a_base_addr.to_le_bytes().map(F::from_u8),
+        (a_base_addr as u64).to_le_bytes().map(F::from_u8),
     );
     tester.write::<RV64_REGISTER_NUM_LIMBS>(
         ptr_as,
         rs2_ptr,
-        b_base_addr.to_le_bytes().map(F::from_u8),
+        (b_base_addr as u64).to_le_bytes().map(F::from_u8),
     );
     tester.write::<RV64_REGISTER_NUM_LIMBS>(
         ptr_as,
         rd_ptr,
-        result_base_addr.to_le_bytes().map(F::from_u8),
+        (result_base_addr as u64).to_le_bytes().map(F::from_u8),
     );
 
     let a_c0_limbs: Vec<F> = biguint_to_limbs_vec(&a_c0, NUM_LIMBS)
